accelerometer_test: set spi master div_coef to 200
The clock is 10MHz and the default div_coef is 10k. This sets the spi
clock to 500Hz (1kHz/spi clk transition). The accelerometer spi read
spec is 10MHz, so we have plenty of room to speed up the reads.

Setting the div_coef to 200 sets the spi clock to 25kHz. It takes 25
iterations to get the new value, so the new response time is about 1ms.